+++ ThreadController.h
@@ -39,7 +39,7 @@ public:
 	bool add(Thread* _thread);
 
 	// remove the thread (given the Thread* or ThreadID)
-	void remove(int _id);
+	void remove(size_t _id);
 	void remove(Thread* _thread);
 
 	// Removes all threads
